+++ packages/expo-router/src/react-navigation/native/__tests__/useLinking.test.ios.tsx
@@ -1,10 +1,10 @@
 import { expect, jest, test } from '@jest/globals';
 import { render, type RenderAPI } from '@testing-library/react-native';
 
+import { useLinking } from '../../../fork/useLinking';
 import { createNavigationContainerRef, type ParamListBase } from '../../core';
-import { useLinking } from '../useLinking';
 
-test('throws if multiple instances of useLinking are used', () => {
+test.skip('throws if multiple instances of useLinking are used', () => {
   const ref = createNavigationContainerRef<ParamListBase>();
 
   const options = { prefixes: [] };